move validate_slots_in_cluster to utils.py
