+++ tests/flow/test_asm.py
@@ -8,28 +8,10 @@ from typing import Optional, Set
 import redis
 
 from includes import Env, VALGRIND, SANITIZER, RUNNER_LABEL
-from utils import migrate_slots_back_and_forth, fill_ts_data, ClusterNode, NUMBER_OF_SLOTS
+from utils import migrate_slots_back_and_forth, fill_ts_data, validate_slots_in_cluster
 
 
 MIGRATION_CYCLES = 10
-
-
-def validate_slots_in_cluster(env):
-    slot_ranges = []
-    for line in env.getConnection(0).execute_command("cluster", "nodes").splitlines():
-        slot_ranges.extend(ClusterNode.from_str(line).slots)
-
-    total = 0
-    min_start = NUMBER_OF_SLOTS
-    max_end = -1
-    for sr in slot_ranges:
-        total += sr.end - sr.start + 1
-        min_start = min(min_start, sr.start)
-        max_end = max(max_end, sr.end)
-
-    assert min_start == 0
-    assert max_end == NUMBER_OF_SLOTS - 1
-    assert total == NUMBER_OF_SLOTS
 
 
 def test_asm_without_data():
